requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-dotenv>=1.0
supabase>=2.5
requests>=2.31
pandas>=2.2
numpy>=1.26
matplotlib>=3.8
pydeck>=0.9
streamlit>=1.32
scikit-learn>=1.4
